celery/worker/requirements.txt
modal-client
Pillow
celery
requests
openai==0.28
supabase